% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSOoptim.R
\name{BSOoptim}
\alias{BSOoptim}
\title{Implementation of the Beetle Swarm Optimization(BSO) algorithm, which is mixture of BAS and PSO.}
\usage{
BSOoptim(fn, init = NULL, constr = NULL, lower = c(-50, -50),
  upper = c(50, 50), n = 300, s = floor(10 + 2 *
  sqrt(length(lower))), w = c(0.9, 0.4), w_vs = 0.4, step = 10,
  step_w = c(0.9, 0.4), c = 8, v = c(-5.12, 5.12), trace = T,
  seed = NULL, pen = 1e+06)
}
\arguments{
\item{fn}{objective function; function need to be optimized}

\item{init}{default = NULL, it will generate randomly; Or you can specify it as a matrix. For example, if your problem
is defined as m beetles and n dimensions, you can specify init as a \eqn{m*n} matrix.}

\item{constr}{constraint function. For example, you can formulate \eqn{x<=10} as
\eqn{constr = function(x) return(x - 10)}.}

\item{lower}{lower of parameters to be estimated}

\item{upper}{upper of parameters}

\item{n}{maximum number of iterations}

\item{s}{a positive integer, beetles number. Default = \eqn{floor(10+2*\sqrt{dimensions})}}

\item{w}{a vector for calculating inertia weight, \eqn{(\omega_{max},\omega_{min})}, default c(0.9,0.4).
the strategy of decreasing inertia weight is as follows.
\deqn{\omega = \omega_{max} - (\omega_{max}-\omega_{min})\frac{k}{n}}.
\eqn{k} is the current number of iteration and \eqn{n} is the maximum number of iterations.}

\item{w_vs}{a positive constant belongs to [0,1]. Default = 0.4.
\deqn{X_{is}^{k+1}=X_{is}^k+ \lambda V_{is}^k+(1-\lambda)\xi_{is}^k}
\emph{w_vs} is \eqn{\lambda}, which means the weight of speed(PSO) and beetle movement(BAS).}

\item{step}{initial step-size of beetles}

\item{step_w}{a vector used for step-size updating, default c(0.9,0.4).
\deqn{\delta_{k+1} = \eta \delta_{k}}
\deqn{\eta = \delta_{w1}(\frac{\delta_{w0}}{\delta_{w1}})^{\frac{1}{1+10*k/n}}} step_w = \eqn{(\delta_{w0},\delta_{w1})}}

\item{c}{ratio of step-size and searching distance.\deqn{d = \frac{step}{c}}}

\item{v}{the speed range of beetles. Default = c(-5.12, 5.12)}

\item{trace}{default = T; trace the process of BAS iteration.}

\item{seed}{random seed; default = NULL}

\item{pen}{penalty conefficient usually predefined as a large enough value, default 1e6}
}
\value{
A list including best beetle position ($par) and corresponding objective function value($value).
}
\description{
\emph{BSO} is the combination of \emph{BAS} and \emph{PSO}. It improves the \emph{BAS} algorithm
by expanding an individual to a group in the way of \emph{PSO}. Note: In a word, 'swarm' in \emph{BSAS} is more like
that a beetle has many antennae pairs(searching direction) while the 'swarm' in \emph{BSO} means group just like PSO.
}
\examples{
#======== examples start =======================
# >>>> example with constraint: Mixed integer nonlinear programming <<<<
pressure_Vessel <- list(
obj = function(x){
  x1 <- floor(x[1])*0.0625
  x2 <- floor(x[2])*0.0625
  x3 <- x[3]
  x4 <- x[4]
  result <- 0.6224*x1*x3*x4 + 1.7781*x2*x3^2 +3.1611*x1^2*x4 + 19.84*x1^2*x3
},
con = function(x){
  x1 <- floor(x[1])*0.0625
  x2 <- floor(x[2])*0.0625
  x3 <- x[3]
  x4 <- x[4]
  c(
    0.0193*x3 - x1,#<=0
    0.00954*x3 - x2,
    750.0*1728.0 - pi*x3^2*x4 - 4/3*pi*x3^3
  )
}
)
result<-
BSOoptim(fn = pressure_Vessel$obj,
         init = NULL,
         constr = pressure_Vessel$con,
         lower = c( 1, 1, 10, 10),
         upper = c(100, 100, 200, 200),
         n = 1000,
         w = c(0.9,0.4),
         w_vs = 0.9,
         step = 100,
         step_w = c(0.9,0.4),
         c = 35,
         v = c(-5.12,5.12),
         trace = F,seed = 1,
         pen = 1e6)
 result$par; result$value
# >>>> example without constraint: Michalewicz function <<<<
mich <- function(x){
y1 <- -sin(x[1])*(sin((x[1]^2)/pi))^20
y2 <- -sin(x[2])*(sin((2*x[2]^2)/pi))^20
return(y1+y2)
}
result <-
 BSOoptim(fn = mich,
           init = NULL,
           lower = c(-6,0),
           upper = c(-1,2),
           n = 100,
           step = 5,
           s = 10,seed = 1, trace = F)
result$par; result$value
#======== examples end =======================
}
\references{
Wang T, Yang L, Liu Q. Beetle Swarm Optimization Algorithm:Theory and Application.2018. arXiv:1808.00206v1.\url{https://arxiv.org/abs/1808.00206}
}
